Updated twitter auth structure
