+++ src/components/Profile/Verification/NameAndIdVerification.js
@@ -1,16 +1,14 @@
-import { useState, useEffect, useContext } from "react";
+import { useState, useContext } from "react";
 import Link from "next/link";
 import { useMoralisCloudFunction } from "react-moralis";
 import PersonaVerification from "./PersonaVerification";
 import VerificationButton from "./VerificationButton";
 import { useMoralis } from "react-moralis";
 import StatusContext from "../../../../store/status-context";
-import LoadingContext from "../../../../store/loading-context";
 
 const NameAndIdVerification = ({ nextStep, isStageNameDifferent, setIsStageNameDifferent, artistStageName, setArtistStageName, personaInquiryIdData }) => {
 	const { user } = useMoralis();
-	const [, setLoading] = useContext(LoadingContext);
-	const [, , setSuccess, setError] = useContext(StatusContext);
+	const [, , , setError] = useContext(StatusContext);
 
 	const [isVerificationModalOpen, setIsVerificationModalOpen] = useState(false);
 	const [artistStageNameSave, setArtistStageNameSave] = useState(false);
